requirements.txt
adbutils==2.9.4
certifi==2025.8.3
charset-normalizer==3.4.3
decorator==5.2.1
deprecation==2.1.0
idna==3.10
lxml==6.0.0
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pytesseract==0.3.13
requests==2.32.4
retry2==0.9.5
uiautomator2==3.4.0
urllib3==2.5.0
